+++ apps/web/src/app/(needLogin)/story_community/_components/CategoryTab/CategoryTab.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useAtom } from 'jotai';
 
 import { storyCommunityPageConfig } from '@/constants';
