+++ apps/web/src/app/(needLogin)/story_community/_components/SearchBar/_components/SearchInput/SearchInput.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useSetAtom } from 'jotai';
 import { Controller, useForm } from 'react-hook-form';
