+++ apps/web/src/app/(needLogin)/story_community/_components/SearchBar/_components/SuggestionModal/SuggestionModal.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useAtom } from 'jotai';
 
 import { mockKeywordsArray } from '@/app/(needLogin)/story_community/_mocks';
