+++ apps/web/src/app/(needLogin)/story_community/_components/Title/index.ts
@@ -1,0 +1,1 @@
+export * from './Title';
