+++ apps/web/src/app/(needLogin)/story_community/_components/index.ts
@@ -1,2 +1,4 @@
 export * from './CategoryTab';
 export * from './SearchBar';
+export * from './StoryList';
+export * from './Title';
